HW03_check_2   initDB.sql - добавлен скрипт создания таблицы meal и уникального индекса по (user_id,datatime)   populateDB.sql - добавлен скрипт инициализации таблицы meal   Meal - добавлены конструктор по умолчанию и сеттеры по полям dateTime,description,calories для возможности работы через JdbcTemplate, добавлен конструктор копирования для использования в тестах обновления (update,updateAlienMeal) чтобы не портить статические данные MealTestData   MealTestData - по аналогии с UserTestData добавлены статические данные аналогично скриптам загрузки meal (populateDb.sql)   JdbcMealRepositoryImpl - реализованы методы взаимодействия(CRUD) с БД через Spring Jdbc Template.   MealServiceTest - реализованы тесты из MealService для JdbcMealRepository, также тесты на получение, удаление и обновления чужой еды (getAlienMeal, deleteAlienMeal, updateAlienMeal).   test/...resources/spring/spring-inmemory.xml - сканирование каталогов по inmemory   InMemoryAdminRestControllerSpringTest, InMemoryAdminRestControllerTest - добавлен в контекст spring-inmemory.xml

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepositoryImpl.java
@@ -12,21 +12,12 @@ import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
 public class JdbcMealRepositoryImpl implements MealRepository {
-
-    private static final RowMapper<Meal> ROW_MAPPER=new RowMapper<Meal>() {
-        @Override
-        public Meal mapRow(ResultSet rs, int i) throws SQLException {
-            return new Meal(rs.getInt("id"),rs.getTimestamp("datetime").toLocalDateTime(),
-                    rs.getString("description"),rs.getInt("calories"));
-        }
-    };
+    private static final RowMapper<Meal> ROW_MAPPER = BeanPropertyRowMapper.newInstance(Meal.class);
 
     private final JdbcTemplate jdbcTemplate;
 
@@ -36,56 +27,57 @@ public class JdbcMealRepositoryImpl implements MealRepository {
 
     @Autowired
     public JdbcMealRepositoryImpl(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
-        this.jdbcTemplate=jdbcTemplate;
+        this.jdbcTemplate = jdbcTemplate;
         this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
-        this.insertMeal=new SimpleJdbcInsert(jdbcTemplate)
+        this.insertMeal = new SimpleJdbcInsert(jdbcTemplate)
                 .withTableName("meal")
                 .usingGeneratedKeyColumns("id");
     }
 
     @Override
     public Meal save(Meal meal, int userId) {
-        MapSqlParameterSource mapSqlSource=new MapSqlParameterSource()
-                .addValue("id",meal.getId())
-                .addValue("user_id",userId)
-                .addValue("datetime",meal.getDateTime())
-                .addValue("description",meal.getDescription())
-                .addValue("calories",meal.getCalories());
-        if(meal.isNew()){
-            Number key=insertMeal.executeAndReturnKey(mapSqlSource);
+        MapSqlParameterSource mapSqlSource = new MapSqlParameterSource()
+                .addValue("id", meal.getId())
+                .addValue("user_id", userId)
+                .addValue("datetime", meal.getDateTime())
+                .addValue("description", meal.getDescription())
+                .addValue("calories", meal.getCalories());
+        if (meal.isNew()) {
+            Number key = insertMeal.executeAndReturnKey(mapSqlSource);
             meal.setId(key.intValue());
-        }
-        else{
-            if(namedParameterJdbcTemplate.update("UPDATE meal SET datetime=:datetime, description=:description," +
-                    "calories=:calories WHERE id=:id",mapSqlSource)==0){
+        } else {
+            if (namedParameterJdbcTemplate.update("UPDATE meal SET datetime=:datetime, description=:description," +
+                    "calories=:calories WHERE id=:id AND user_id=:user_id", mapSqlSource) == 0) {
                 return null;
-            };
+            }
+            ;
         }
         return meal;
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update("DELETE FROM meal WHERE id=?",id)!=0;
+        return jdbcTemplate.update("DELETE FROM meal WHERE id=? AND user_id=?", id, userId) != 0;
     }
 
     @Override
     public Meal get(int id, int userId) {
-        List<Meal> meals=jdbcTemplate.query("SELECT * FROM meal WHERE id=?",ROW_MAPPER,id);
+        List<Meal> meals = jdbcTemplate.query("SELECT * FROM meal WHERE id=? AND user_id=?", ROW_MAPPER, id, userId);
         return DataAccessUtils.singleResult(meals);
     }
 
     @Override
     public List<Meal> getAll(int userId) {
-        return jdbcTemplate.query("SELECT * FROM meal  WHERE user_id=? ORDER BY datetime desc",ROW_MAPPER,userId);
+        return jdbcTemplate.query("SELECT * FROM meal  WHERE user_id=? ORDER BY datetime desc", ROW_MAPPER, userId);
     }
 
     @Override
     public List<Meal> getBetween(LocalDateTime startDate, LocalDateTime endDate, int userId) {
-        MapSqlParameterSource mapSqlSource=new MapSqlParameterSource()
-                .addValue("user_id",userId)
-                .addValue("startDate",startDate)
-                .addValue("endDate",endDate);
-        return namedParameterJdbcTemplate.query("SELECT * FROM meal WHERE user_id=:user_id AND datetime BETWEEN :startDate AND :endDate",mapSqlSource,ROW_MAPPER);
+        MapSqlParameterSource mapSqlSource = new MapSqlParameterSource()
+                .addValue("user_id", userId)
+                .addValue("startDate", startDate)
+                .addValue("endDate", endDate);
+        return namedParameterJdbcTemplate.query("SELECT * FROM meal WHERE user_id=:user_id AND datetime BETWEEN :startDate AND :endDate ORDER BY datetime desc", mapSqlSource, ROW_MAPPER);
+        //return jdbcTemplate.query("SELECT * FROM meal WHERE user_id=? AND datetime BETWEEN ? AND ? ORDER BY datetime DESC",ROW_MAPPER,userId,startDate,endDate);
     }
 }
